requirements.txt
requests==2.32.3
python-dotenv==1.0.1

# package
setuptools==75.6.0